reorganize Rainbow definition into multiple smaller files
