+++ chromatic/rainbows/rainbow.py
@@ -1,6 +1,5 @@
 from ..imports import *
 from ..talker import Talker
-from ..resampling import *
 from .readers import *
 from .writers import *
 
@@ -459,817 +458,17 @@ class Rainbow(Talker):
         n = self.__class__.__name__.replace("Rainbow", "🌈")
         return f"<{n}({self.nwave}w, {self.ntime}t)>"
 
-    def __add__(self, object):
-        """
-        Add the flux of a rainbow and an input array (or another rainbow)
-        and output in a new rainbow object.
-        Currently only supports flux addition.
-
-        Parameters
-        ----------
-        object : Array or float.
-            Multiple options:
-            1) float
-            2) 1D array with same length as wavelength axis
-            3) 1D array with same length as time axis
-            4) 2D array with same shape as rainbow flux
-            5) Rainbow object with same dimensions as self.
-
-        Returns
-        ----------
-        rainbow : Rainbow() with the same parameters as self but with added
-        flux.
-        """
-
-        # Create new Rainbow() to store results in.
-        result = copy.deepcopy(self)
-
-        try:  # Object is rainbow
-            if np.array_equal(
-                self.wavelike["wavelength"], object.wavelike["wavelength"]
-            ) and np.array_equal(self.timelike["time"], object.timelike["time"]):
-                result.fluxlike["flux"] += object.fluxlike["flux"]
-            else:
-                print("Objects do not share wavelength/time axes")
-                return
-
-        except AttributeError:  # Object is not rainbow
-
-            if type(object) == float or type(object) == int:  # Float or integer.
-                result.fluxlike["flux"] += object
-
-            elif self.shape == object.shape:  # fluxlike
-                result.fluxlike["flux"] += object
-
-            elif len(object) == len(
-                self.wavelike["wavelength"]
-            ):  # Add along wavelength axis
-
-                result.fluxlike["flux"] += np.transpose([object] * self.shape[1])
-                if self.nwave == self.ntime:
-                    raise RuntimeError(
-                        f"{self} has same number of wavelengths and times; we can't tell which is which."
-                    )
-
-            elif len(object) == len(self.timelike["time"]):  # Add along time axis.
-                result.fluxlike["flux"] += np.tile(object, (self.shape[0], 1))
-
-            else:
-                print("Invalid shape in addition: " + str(object.shape))
-                return
-        return result
-
-    def __sub__(self, object):
-        """
-        Subtract the flux of a rainbow from an input array (or another rainbow)
-        and output in a new rainbow object.
-        Currently only supports flux subtraction.
-
-        Parameters
-        ----------
-        object : Array or float.
-            Multiple options:
-            1) float
-            2) 1D array with same length as wavelength axis
-            3) 1D array with same length as time axis
-            4) 2D array with same shape as rainbow flux
-            5) Rainbow object with same dimensions as self.
-
-        Returns
-        ----------
-        rainbow : Rainbow() with the same parameters as self but with subtracted
-        flux.
-        """
-
-        # Create new Rainbow() to store results in.
-        result = copy.deepcopy(self)
-
-        try:  # Object is rainbow.
-            if np.array_equal(
-                self.wavelike["wavelength"], object.wavelike["wavelength"]
-            ) and np.array_equal(self.timelike["time"], object.timelike["time"]):
-                result.fluxlike["flux"] -= object.fluxlike["flux"]
-            else:
-                print("Objects do not share wavelength/time axes")
-                return
-
-        except AttributeError:
-
-            if type(object) == float or type(object) == int:  # Float or integer.
-                result.fluxlike["flux"] -= object
-
-            elif self.shape == object.shape:  # fluxlike
-                result.fluxlike["flux"] -= object
-
-            elif len(object) == len(
-                self.wavelike["wavelength"]
-            ):  # Add along wavelength axis
-                result.fluxlike["flux"] -= np.transpose([object] * self.shape[1])
-                if self.nwave == self.ntime:
-                    raise RuntimeError(
-                        f"{self} has same number of wavelengths and times; we can't tell which is which."
-                    )
-
-            elif len(object) == len(self.timelike["time"]):  # Add along time axis.
-                result.fluxlike["flux"] -= np.tile(object, (self.shape[0], 1))
-
-            else:
-                print("Invalid shape in subtraction: " + str(object.shape))
-                return
-        return result
-
-    def __mul__(self, object):
-        """
-        Multiply the flux of a rainbow and an input array (or another rainbow)
-        and output in a new rainbow object.
-        Currently only supports flux multiplication.
-
-        Parameters
-        ----------
-        object : Array or float.
-            Multiple options:
-            1) float
-            2) 1D array with same length as wavelength axis
-            3) 1D array with same length as time axis
-            4) 2D array with same shape as rainbow flux
-            5) Rainbow object with same dimensions as self.
-
-        Returns
-        ----------
-        rainbow : Rainbow() with the same parameters as self but with multiplied
-        flux.
-        """
-
-        # Create new Rainbow() to store results in.
-        result = copy.deepcopy(self)
-
-        try:  # Object is rainbow
-            if np.array_equal(
-                self.wavelike["wavelength"], object.wavelike["wavelength"]
-            ) and np.array_equal(self.timelike["time"], object.timelike["time"]):
-                result.fluxlike["flux"] *= object.fluxlike["flux"]
-            else:
-                print("Objects do not share wavelength/time axes")
-                return
-
-        except AttributeError:
-
-            if type(object) == float or type(object) == int:  # Float or integer.
-                result.fluxlike["flux"] *= object
-
-            elif self.shape == object.shape:  # fluxlike
-                result.fluxlike["flux"] *= object
-
-            elif len(object) == len(
-                self.wavelike["wavelength"]
-            ):  # Add along wavelength axis
-                result.fluxlike["flux"] *= np.transpose([object] * self.shape[1])
-                if self.nwave == self.ntime:
-                    raise RuntimeError(
-                        f"{self} has same number of wavelengths and times; we can't tell which is which."
-                    )
-
-            elif len(object) == len(self.timelike["time"]):  # Add along time axis.
-                result.fluxlike["flux"] *= np.tile(object, (self.shape[0], 1))
-
-            else:
-                print("Invalid shape in multiplication: " + str(object.shape))
-                return
-        return result
-
-    def __truediv__(self, object):
-        """
-        Divide the flux of a rainbow and an input array (or another rainbow)
-        and output in a new rainbow object.
-        Currently only supports flux division.
-
-        Parameters
-        ----------
-        object : Array or float.
-            Multiple options:
-            1) float
-            2) 1D array with same length as wavelength axis
-            3) 1D array with same length as time axis
-            4) 2D array with same shape as rainbow flux
-            5) Rainbow object with same dimensions as self.
-
-        Returns
-        ----------
-        rainbow : Rainbow() with the same parameters as self but with divided
-        flux.
-        """
-
-        # Create new Rainbow() to store results in.
-        result = copy.deepcopy(self)
-
-        try:  # Object is another rainbow.
-            if np.array_equal(
-                self.wavelike["wavelength"], object.wavelike["wavelength"]
-            ) and np.array_equal(self.timelike["time"], object.timelike["time"]):
-                result.fluxlike["flux"] /= object.fluxlike["flux"]
-            else:
-                print("Objects do not share wavelength/time axes")
-                return
-
-        except AttributeError:
-
-            if type(object) == float or type(object) == int:  # float
-                result.fluxlike["flux"] /= object
-
-            elif self.shape == object.shape:  # fluxlike array
-                result.fluxlike["flux"] /= object
-
-            elif len(object) == len(
-                self.wavelike["wavelength"]
-            ):  # Add along wavelength axis
-                result.fluxlike["flux"] /= np.transpose([object] * self.shape[1])
-                if self.nwave == self.ntime:
-                    raise RuntimeError(
-                        f"{self} has same number of wavelengths and times; we can't tell which is which."
-                    )
-            elif len(object) == len(self.timelike["time"]):  # Add along time axis.
-                result.fluxlike["flux"] /= np.tile(object, (self.shape[0], 1))
-
-            else:
-                print("Invalid shape in division: " + str(object.shape))
-                return
-        return result
-
-    def __eq__(self, other):
-        """
-        Test whether (self) == (other).
-        """
-        # start by assuming the Rainbows are identical
-        same = True
-
-        # loop through the core dictionaries
-        for d in self._core_dictionaries:
-
-            # pull out each core dictionary from both
-            d1, d2 = vars(self)[d], vars(other)[d]
-
-            # loop through elements of each dictionary
-            for k in d1:
-
-                # test that all elements match for both
-                same *= np.all(d1[k] == d2.get(k, None))
-
-        return same
-
-    def normalize(self):
-        """
-        Normalize by dividing through by the median spectrum.
-        """
-
-        # TODO, think about more careful treatment of uncertainties + good/bad data
-        return self / np.nanmedian(self.flux, axis=self.timeaxis)
-
-    def bin(self, dt=None, time=None, R=None, dw=None, wavelength=None):
-        """
-        Bin the rainbow in wavelength and/or time.
-
-        The time-setting order of precendence is
-        [`time`, `dt`], meaning that if `time` is set,
-        any values given for `dt` will be ignored.
-        The wavelength-setting order of precendence is
-        [`wavelength`, `dw`, `R`], meaning that if `wavelength`
-        is set any values of `dw` or `R` will be ignored, and
-        if `dw` is set any value of `R` will be ignored.
-
-        Parameters
-        ----------
-        dt : astropy.units.Quantity
-            The d(time) bin size for creating a grid
-            that is uniform in linear space.
-        time : array of astropy.units.Quantity
-            An array of times, if you just want to give
-            it an entirely custom array.
-        R : float
-            The spectral resolution for creating a grid
-            that is uniform in logarithmic space.
-        dw : astropy.units.Quantity
-            The d(wavelength) bin size for creating a grid
-            that is uniform in linear space.
-        wavelength : array of astropy.units.Quantity
-            An array of wavelengths, if you just want to give
-            it an entirely custom array.
-        """
-
-        # self.speak(f'binning')
-
-        # bin first in time
-        binned_in_time = self.bin_in_time(dt=dt, time=time)
-
-        # then bin in wavelength
-        binned = binned_in_time.bin_in_wavelength(R=R, dw=dw, wavelength=wavelength)
-
-        # return the binned object
-        return binned
-
-    def bin_in_time(self, dt=None, time=None):
-        """
-        Bin the rainbow in time.
-
-        Parameters
-        ----------
-        dt : astropy.units.Quantity
-            The d(time) bin size for creating a grid
-            that is uniform in linear space.
-        time : array of astropy.units.Quantity
-            An array of times, if you just want to give
-            it an entirely custom array.
-
-        The time-setting order of precendence is:
-            1) time
-            2) dt
-        """
-
-        # if no bin information is provided, don't bin
-        if (time is None) and (dt is None):
-            return self
-
-        # set up binning parameters
-        binkw = dict(weighting="inversevariance", drop_nans=False)
-        # TODO: make sure drop_nans doesn't skip rows or columns unexpectedly
-        if time is not None:
-            binkw["newx"] = time
-            # self.speak(f'binning to time={time}')
-        elif dt is not None:
-            binkw["dx"] = dt
-            # self.speak(f'binning to dt={dt}')
-
-        # create a new, empty Rainbow
-        new = Rainbow()
-
-        # populate the wavelength information
-        new.wavelike = {**self.wavelike}
-        new.metadata["wscale"] = self.wscale
-
-        # bin the time-like variables
-        # TODO (add more careful treatment of uncertainty + DQ)
-        new.timelike = {}
-        for k in self.timelike:
-            bt, bv = bintogrid(x=self.time, y=self.timelike[k], unc=None, **binkw)
-            new.timelike[k] = bv
-        new.timelike[k] = bt
-
-        # bin the flux-like variables
-        # TODO (add more careful treatment of uncertainty + DQ)
-        # TODO (think about cleverer bintogrid for 2D arrays)
-        new.fluxlike = {}
-        for k in self.fluxlike:
-            # self.speak(f" binning '{k}' in time")
-            # self.speak(f"  original shape was {np.shape(self.fluxlike[k])}")
-
-            if k == "uncertainty":
-                warnings.warn(
-                    """
-                Uncertainties aren't being handled well yet...
-                """
-                )
-                continue
-
-            for w in range(new.nwave):
-                if self.uncertainty is None:
-                    bt, bv = bintogrid(
-                        x=self.time, y=self.fluxlike[k][w, :], unc=None, **binkw
-                    )
-                    bu = None
-                else:
-                    bt, bv, bu = bintogrid(
-                        x=self.time,
-                        y=self.fluxlike[k][w, :],
-                        unc=self.uncertainty[w, :],
-                        **binkw,
-                    )
-
-                if k not in new.fluxlike:
-                    new_shape = (new.nwave, new.ntime)
-                    new.fluxlike[k] = np.zeros(new_shape)
-                    # TODO make this more robust to units
-
-                if k == "uncertainty":
-                    new.fluxlike[k][w, :] = bu
-                else:
-                    new.fluxlike[k][w, :] = bv
-            # self.speak(f"  new shape is {np.shape(new.fluxlike[k])}")
-        new._validate_core_dictionaries()
-        return new
-
-    def bin_in_wavelength(self, R=None, dw=None, wavelength=None):
-        """
-        Bin the rainbow in wavelength.
-
-        Parameters
-        ----------
-        R : float
-            The spectral resolution for creating a grid
-            that is uniform in logarithmic space.
-        dw : astropy.units.Quantity
-            The d(wavelength) bin size for creating a grid
-            that is uniform in linear space.
-        wavelength : array of astropy.units.Quantity
-            An array of wavelengths, if you just want to give
-            it an entirely custom array.
-
-        The wavelength-setting order of precendence is:
-            1) wavelength
-            2) dw
-            3) R
-        """
-
-        # if no bin information is provided, don't bin
-        if (wavelength is None) and (dw is None) and (R is None):
-            return self
-
-        # set up binning parameters
-        binkw = dict(weighting="inversevariance", drop_nans=False)
-        if wavelength is not None:
-            binning_function = bintogrid
-            binkw["newx"] = wavelength
-            wscale = "?"
-            # self.speak(f'binning to wavelength={wavelength}')
-        elif dw is not None:
-            binning_function = bintogrid
-            binkw["dx"] = dw
-            wscale = "linear"
-            # self.speak(f'binning to dw={dw}')
-        elif R is not None:
-            binning_function = bintoR
-            binkw["R"] = R
-            wscale = "log"
-            # self.speak(f'binning to R={R}')
-
-        # create a new, empty Rainbow
-        new = Rainbow()
-
-        # populate the time information
-        new.timelike = {**self.timelike}
-
-        # bin the time-like variables
-        # TODO (add more careful treatment of uncertainty + DQ)
-        new.wavelike = {}
-        for k in self.wavelike:
-            bt, bv = binning_function(
-                x=self.wavelength, y=self.wavelike[k], unc=None, **binkw
-            )
-            new.wavelike[k] = bv
-        new.wavelike[k] = bt
-
-        # bin the flux-like variables
-        # TODO (add more careful treatment of uncertainty + DQ)
-        # TODO (think about cleverer bintogrid for 2D arrays)
-        new.fluxlike = {}
-        for k in self.fluxlike:
-            # self.speak(f" binning '{k}' in wavelength")
-            # self.speak(f"  original shape was {np.shape(self.fluxlike[k])}")
-            if k == "uncertainty":
-                warnings.warn(
-                    """
-                Uncertainties aren't being handled well yet...
-                """
-                )
-                continue
-
-            for t in range(new.ntime):
-                if self.uncertainty is None:
-                    bt, bv = binning_function(
-                        x=self.wavelength, y=self.fluxlike[k][:, t], unc=None, **binkw
-                    )
-                    bu = None
-                else:
-                    bt, bv, bu = binning_function(
-                        x=self.wavelength,
-                        y=self.fluxlike[k][:, t],
-                        unc=self.uncertainty[:, t],
-                        **binkw,
-                    )
-
-                if k not in new.fluxlike:
-                    new_shape = (new.nwave, new.ntime)
-                    new.fluxlike[k] = np.zeros(new_shape)
-                    # TODO make this more robust to units
-
-                if k == "uncertainty":
-                    new.fluxlike[k][:, t] = bu
-                else:
-                    new.fluxlike[k][:, t] = bv
-            # self.speak(f"  new shape is {np.shape(new.fluxlike[k])}")
-        new.metadata["wscale"] = wscale
-        return new
-
-    def imshow(
-        self,
-        ax=None,
-        quantity="flux",
-        w_unit="micron",
-        t_unit="hour",
-        aspect="auto",
-        colorbar=True,
-        origin="upper",
-        **kw,
-    ):
-        """
-        imshow flux as a function of time (x = time, y = wavelength, color = flux).
-
-        Parameters
-        ----------
-        ax : matplotlib.axes.Axes
-            The axes into which to make this plot.
-        w_unit : str, astropy.unit.Unit
-            The unit for plotting wavelengths.
-        t_unit : str, astropy.unit.Unit
-            The unit for plotting times.
-
-        """
-
-        # self.speak(f'imshowing')
-        if ax is None:
-            ax = plt.subplot()
-
-        w_unit, t_unit = u.Unit(w_unit), u.Unit(t_unit)
-
-        if self.wscale == "linear":
-            extent = [
-                (min(self.time) / t_unit).decompose(),
-                (max(self.time) / t_unit).decompose(),
-                (max(self.wavelength) / w_unit).decompose(),
-                (min(self.wavelength) / w_unit).decompose(),
-            ]
-        elif self.wscale == "log":
-            extent = [
-                (min(self.time) / t_unit).decompose(),
-                (max(self.time) / t_unit).decompose(),
-                np.log10(max(self.wavelength) / w_unit),
-                np.log10(min(self.wavelength) / w_unit),
-            ]
-        else:
-            raise RuntimeError("Can't imshow without knowing wscale.")
-
-        with quantity_support():
-            plt.sca(ax)
-            plt.imshow(
-                self.fluxlike[quantity],
-                extent=extent,
-                aspect=aspect,
-                origin=origin,
-                **kw,
-            )
-            if self.wscale == "linear":
-                plt.ylabel(f"Wavelength ({w_unit.to_string('latex_inline')})")
-            elif self.wscale == "log":
-                plt.ylabel(
-                    r"log$_{10}$" + f"[Wavelength/({w_unit.to_string('latex_inline')})]"
-                )
-            plt.xlabel(f"Time ({t_unit.to_string('latex_inline')})")
-            if colorbar:
-                plt.colorbar(ax=ax)
-        return ax
-
-    def plot(self, ax=None, spacing=None, plotkw={}, fontkw={}):
-        """
-        Plot
-
-        Returns
-        -------
-        None.
-
-        """
-        from chromatic import viz
-
-        viz.wavelength_plot(
-            self.flux,
-            self.time,
-            self.wavelength,
-            step_size=spacing,
-            ax=ax,
-            plotkw={},
-            fontkw={},
-        )
-
-    def _setup_animated_plot(self, ax=None, figsize=None, plotkw={}, textkw={}):
-        """
-        Wrapper to set up the basics of animate-able plot.
-
-        This works for any general plot that has a single-color
-        line or set of points (using `plt.plot`), with a text
-        label in the upper right corner.
-
-        Parameters
-        ----------
-        ax : matplotlib.axes.Axes
-            The axes into which the plot should be drawn.
-            If None, a new one will be created.
-        figsize : tuple
-            (width, height) of the figure, if one needs
-            to be created (= if ax isn't specified)
-        plotkw : dict
-            A dictionary of keywords to be passed to `plt.plot`
-        textkw : dict
-            A dictionary of keywords to be passed to `plt.text`
-        """
-
-        # make sure the ax and figure are defined
-        if ax is None:
-            fig, ax = plt.subplots(figsize=figsize)
-        else:
-            fig = ax.get_figure()
-
-        default_plotkw = dict()
-        plot = plt.plot([], [], **default_plotkw, **plotkw)[0]
-
-        default_textkw = dict(
-            x=0.98, y=0.96, s="", ha="right", va="top", transform=ax.transAxes
-        )
-        text = plt.text(**default_textkw, **textkw)
-
-        # return a dictionary with things that will be useful to hang onto
-        return dict(fig=fig, ax=ax, plot=plot, text=text)
-
-    def animate_lightcurves(
-        self,
-        filename="animated-lightcurves.gif",
-        fps=5,
-        figsize=None,
-        xlim=[None, None],
-        ylim=[None, None],
-        plotkw={},
-        textkw={},
-    ):
-        """
-        Create an animation that shows how the lightcurve changes
-        as we flip through every wavelength.
-
-        Parameters
-        ----------
-        filename : str
-            Name of file you'd like to save results in.
-            Currently supports only .gif files.
-        fps : float
-            frames/second of animation
-        figsize : tuple
-            (width, height) of the figure
-        xlim : tuple
-            Custom xlimits for the plot
-        ylim : tuple
-            Custom ylimits for the plot
-        plotkw : dict
-            A dictionary of keywords to be passed to `plt.plot`
-        textkw : dict
-            A dictionary of keywords to be passed to `plt.text`
-        """
-
-        with quantity_support():
-
-            # keep track of the things needed for the animation
-            self._animate_lightcurves_components = self._setup_animated_plot(
-                figsize=figsize, plotkw=plotkw, textkw=textkw
-            )
-
-            ax = self._animate_lightcurves_components["ax"]
-
-            # set the plot limits
-            ax.set_xlim(xlim[0] or np.min(self.time), xlim[1] or np.max(self.time))
-            ax.set_ylim(
-                ylim[0] or 0.995 * np.min(self.flux),
-                ylim[1] or 1.005 * np.max(self.flux),
-            )
-            # set the axis labels
-            ax.set_xlabel(f"Time ({self.time.unit.to_string('latex_inline')})")
-            ax.set_ylabel(f"Relative Flux")
-
-            def update(frame):
-                """
-                This function will be called to update each frame
-                of the animation.
-
-                Parameters
-                ----------
-                frame : int
-                    An integer that will advance with each frame.
-                """
-
-                # pull out the x and y values to plot
-                x = self.time
-                y = self.flux[frame]
-
-                # update the label in the corner
-                self._animate_lightcurves_components["text"].set_text(
-                    f"w = {self.wavelength[frame].value:0.2f} {self.wavelength.unit.to_string('latex')}"
-                )
-
-                # update the plot data
-                self._animate_lightcurves_components["plot"].set_data(x, y)
-
-                return (
-                    self._animate_lightcurves_components["text"],
-                    self._animate_lightcurves_components["plot"],
-                )
-
-            # hold onto this update function in case we need it elsewhere
-            self._animate_lightcurves_components["update"] = update
-
-            # make and save the animation
-            ani = matplotlib.animation.FuncAnimation(
-                self._animate_lightcurves_components["fig"],
-                update,
-                frames=np.arange(0, self.nwave),
-                blit=True,
-            )
-            ani.save(filename, fps=fps)
-
-    def animate_spectra(
-        self,
-        filename="animated-spectra.gif",
-        fps=5,
-        figsize=None,
-        xlim=[None, None],
-        ylim=[None, None],
-        plotkw={},
-        textkw={},
-    ):
-        """
-        Create an animation that shows how the lightcurve changes
-        as we flip through every wavelength.
-
-        Parameters
-        ----------
-        filename : str
-            Name of file you'd like to save results in.
-            Currently supports only .gif files.
-        fps : float
-            frames/second of animation
-        figsize : tuple
-            (width, height) of the figure
-        xlim : tuple
-            Custom xlimits for the plot
-        ylim : tuple
-            Custom ylimits for the plot
-        plotkw : dict
-            A dictionary of keywords to be passed to `plt.plot`
-        textkw : dict
-            A dictionary of keywords to be passed to `plt.text`
-        """
-
-        with quantity_support():
-
-            # keep track of the things needed for the animation
-            self._animate_spectra_components = self._setup_animated_plot(
-                figsize=figsize, plotkw=plotkw, textkw=textkw
-            )
-
-            ax = self._animate_spectra_components["ax"]
-
-            # set the plot limits
-            ax.set_xlim(
-                xlim[0] or np.min(self.wavelength), xlim[1] or np.max(self.wavelength)
-            )
-            ax.set_ylim(
-                ylim[0] or 0.995 * np.min(self.flux),
-                ylim[1] or 1.005 * np.max(self.flux),
-            )
-            # set the axis labels
-            ax.set_xlabel(
-                f"Wavelength ({self.wavelength.unit.to_string('latex_inline')})"
-            )
-            ax.set_ylabel(f"Relative Flux")
-
-            def update(frame):
-                """
-                This function will be called to update each frame
-                of the animation.
-
-                Parameters
-                ----------
-                frame : int
-                    An integer that will advance with each frame.
-                """
-
-                # pull out the x and y values to plot
-                x = self.wavelength
-                y = self.flux[:, frame]
-
-                # update the label in the corner
-                self._animate_spectra_components["text"].set_text(
-                    f"t = {self.time[frame].value:0.2f} {self.time.unit.to_string('latex')}"
-                )
-
-                # update the plot data
-                self._animate_spectra_components["plot"].set_data(x, y)
-
-                return (
-                    self._animate_spectra_components["text"],
-                    self._animate_spectra_components["plot"],
-                )
-
-            # hold onto this update function in case we need it elsewhere
-            self._animate_spectra_components["update"] = update
-
-            # make and save the animation
-            ani = matplotlib.animation.FuncAnimation(
-                self._animate_spectra_components["fig"],
-                update,
-                frames=np.arange(0, self.ntime),
-                blit=True,
-            )
-            ani.save(filename, fps=fps)
+    # import the basic operations for Rainbows
+    from .operations import __add__, __sub__, __mul__, __truediv__, __eq__
+
+    # import other axtions that return other Rainbows
+    from .actions import normalize, bin, bin_in_time, bin_in_wavelength
+
+    # import visualizations that can act on Rainbows
+    from .visualizations import (
+        imshow,
+        plot,
+        animate_lightcurves,
+        animate_spectra,
+        _setup_animated_plot,
+    )
