+++ chromatic/tests/__init__.py
@@ -1,3 +1,6 @@
-from .test_rainbow import *
-from .test_simulated_rainbow import *
-from .test_visualizations import *
+from .test_io import *
+from .test_rainbow_actions import *
+from .test_rainbow_basics import *
+from .test_rainbow_operations import *
+from .test_rainbow_simulations import *
+from .test_rainbow_visualizations import *
