+++ chromatic/tests/test_rainbow_basics.py
@@ -1,0 +1,17 @@
+from ..rainbows import *
+from .setup_tests import *
+
+
+def test_basic_rainbow():
+
+    nw, nt = 20, 40
+    r = Rainbow(
+        wavelength=np.linspace(0.5, 5, nw) * u.micron,
+        time=np.linspace(-1, 1, nt) * u.hour,
+        flux=np.ones((nw, nt)),
+    )
+
+    assert r.shape == (nw, nt)
+    assert r.nwave == nw
+    assert r.ntime == nt
+    assert r.nflux == nw * nt
